requirements.txt
streamlit
pandas
tabula-py
pdfplumber
pdf2image
pytesseract
